refactor: move event.dart into new event folder, split out parts dependent on flutter/material.dart
