+++ app/lib/domain/models/event/event_list_model.dart
@@ -1,0 +1,38 @@
+import 'dart:collection';
+
+import 'package:flutter/foundation.dart';
+
+import 'event.dart';
+
+/// Storage for all the events that units will respond to.
+class EventListModel extends ChangeNotifier {
+  EventListModel({required List<Event> events}) {
+    _events.addAll(events);
+  }
+
+  EventListModel.blank();
+
+  /// Internal, private state of the event list.
+  final List<Event> _events = [];
+
+  /// An unmodifiable view of the events in the event list.
+  UnmodifiableListView<Event> get events => UnmodifiableListView(_events);
+
+  /// The current total number of events.
+  int get numEvents => _events.length;
+
+  /// Adds [event] to event list. This and [removeAll] are the only ways to
+  /// modify the event list from the outside.
+  void add(Event event) {
+    _events.add(event);
+    // This call tells the widgets that are listening to this model to rebuild.
+    notifyListeners();
+  }
+
+  /// Removes all events from the event list.
+  void removeAll() {
+    _events.clear();
+    // This call tells the widgets that are listening to this model to rebuild.
+    notifyListeners();
+  }
+}
