+++ app/lib/domain/models/unit/unit.dart
@@ -1,6 +1,5 @@
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-
-import '../../../map.dart';
+// import 'package:google_maps_flutter/google_maps_flutter.dart' as maps;
+import 'package:latlng/latlng.dart';
 
 /// An ambulance vehicle, air ambulance or other resource.
 class Unit {
@@ -13,27 +12,30 @@ class Unit {
   /// A unique identifier for this resource.
   final String callsign;
 
-  /// Path to the image used as this [Event]'s icon.
-  String get _iconAsset => switch (vehicleType) {
-    VehicleType.dca => 'assets/images/dca.png',
-    VehicleType.rrv => 'assets/images/rrv.png',
-    VehicleType.helicopter => 'assets/images/tvaa.jpg',
-    VehicleType.criticalCareCar => 'assets/images/hems-car.png',
-  };
+  // /// Path to the image used as this [Event]'s icon.
+  // String get _iconAsset => switch (vehicleType) {
+  //   VehicleType.dca => 'assets/images/dca.png',
+  //   VehicleType.rrv => 'assets/images/rrv.png',
+  //   VehicleType.helicopter => 'assets/images/tvaa.jpg',
+  //   VehicleType.criticalCareCar => 'assets/images/hems-car.png',
+  // };
 
   /// The current latitude and longitude of this unit.
   LatLng location;
 
-  /// Gets a [Marker] for showing this [Unit] on a map.
-  Marker get mapMarker => Marker(
-    markerId: MarkerId(callsign),
-    position: LatLng(location.latitude, location.longitude),
-    icon: markerIcon(_iconAsset),
-    infoWindow: InfoWindow(
-      title: '$callsign (${vehicleType.name})',
-      snippet: location.toString(),
-    ),
-  );
+  // /// Gets a [Marker] for showing this [Unit] on a map.
+  // maps.Marker get mapMarker => maps.Marker(
+  //   markerId: maps.MarkerId(callsign),
+  //   position: maps.LatLng(
+  //     location.latitude.degrees,
+  //     location.longitude.degrees,
+  //   ),
+  //   icon: markerIcon(_iconAsset),
+  //   infoWindow: maps.InfoWindow(
+  //     title: '$callsign (${vehicleType.name})',
+  //     snippet: location.toString(),
+  //   ),
+  // );
 
   /// The type of vehicle that this is.
   final VehicleType vehicleType;
