+++ app/lib/main.dart
@@ -3,10 +3,7 @@ import 'package:provider/provider.dart';
 
 import 'api/env.dart';
 import 'config/dependencies.dart';
-// import 'data/services/local_data_service.dart';
-import 'domain/models/event.dart';
-// import 'domain/models/unit.dart';
-// import 'map.dart';
+import 'domain/models/event/event_list_model.dart';
 import 'routing/router.dart';
 
 void main() async {
